Refactoring Activity Rate Functions in the Issue App
It has been refactored for date validation functions, to calculate the metrics of the Activity Rate and to pick up all issues
